# 修复 - 载入编辑器历史ID为空时会产生错误的问题 - 食物动画编辑页面的食物在动画中显示异常的问题

+++ VPet.ModMaker/ViewModels/ModMakerWindowVM.cs
@@ -125,6 +125,7 @@ public class ModMakerWindowVM : ObservableObjectX<ModMakerWindowVM>
         {
             if (LPSConvert.DeserializeObject<ModMakeHistory>(line) is not ModMakeHistory history)
                 continue;
+            history.ID ??= string.Empty;
             set.Add(history);
         }
         Histories.AddRange(set.OrderByDescending(h => h.LastTime));
@@ -181,8 +182,8 @@ public class ModMakerWindowVM : ObservableObjectX<ModMakerWindowVM>
         )
             return;
         Histories.Clear();
-        Histories.Clear();
-        File.WriteAllText(ModMakerInfo.HistoryFile, string.Empty);
+        if (File.Exists(ModMakerInfo.HistoryFile))
+            File.WriteAllText(ModMakerInfo.HistoryFile, string.Empty);
     }
     #endregion
